+++ tapis-appslib/src/main/java/edu/utexas/tacc/tapis/apps/model/FileInput.java
@@ -1,7 +1,9 @@
 package edu.utexas.tacc.tapis.apps.model;
 
+import com.google.gson.JsonObject;
 import edu.utexas.tacc.tapis.shared.utils.TapisUtils;
 import edu.utexas.tacc.tapis.apps.model.App.FileInputMode;
+import static edu.utexas.tacc.tapis.apps.model.App.DEFAULT_NOTES;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -32,6 +34,7 @@ public final class FileInput
   private final String description;
   private final FileInputMode inputMode;
   private final Boolean autoMountLocal;
+  private final JsonObject notes; // metadata as json
   private final String sourceUrl;
   private final String targetPath;
 
@@ -46,17 +49,19 @@ public final class FileInput
     description = null;
     inputMode = DEFAULT_INPUT_MODE;
     autoMountLocal = DEFAULT_AUTO_MOUNT_LOCAL;
+    notes = DEFAULT_NOTES;
     sourceUrl = null;
     targetPath = null;
   }
 
   public FileInput(String name1, String description1, FileInputMode inputMode1, Boolean autoMountLocal1,
-                   String sourceUrl1, String targetPath1)
+                   JsonObject notes1, String sourceUrl1, String targetPath1)
   {
     name = name1;
     description = description1;
     inputMode = (inputMode1 == null) ? DEFAULT_INPUT_MODE : inputMode1;
     autoMountLocal = (autoMountLocal1 == null) ? DEFAULT_AUTO_MOUNT_LOCAL : autoMountLocal1;
+    notes = (notes1 == null) ? DEFAULT_NOTES : notes1;
     sourceUrl = sourceUrl1;
     targetPath = targetPath1;
   }
@@ -68,6 +73,7 @@ public final class FileInput
   public String getDescription() { return description; }
   public FileInputMode getInputMode() { return inputMode; }
   public Boolean isAutoMountLocal() { return autoMountLocal; }
+  public JsonObject getNotes() { return notes; }
   public String getSourceUrl() { return sourceUrl; }
   public String getTargetPath() { return targetPath; }