+++ tapis-appslib/src/main/java/edu/utexas/tacc/tapis/apps/model/FileInputArray.java
@@ -1,5 +1,6 @@
 package edu.utexas.tacc.tapis.apps.model;
 
+import com.google.gson.JsonObject;
 import edu.utexas.tacc.tapis.apps.model.App.FileInputMode;
 import edu.utexas.tacc.tapis.shared.utils.TapisUtils;
 import org.slf4j.Logger;
@@ -8,6 +9,7 @@ import org.slf4j.LoggerFactory;
 import java.util.ArrayList;
 import java.util.List;
 
+import static edu.utexas.tacc.tapis.apps.model.App.DEFAULT_NOTES;
 import static edu.utexas.tacc.tapis.apps.model.FileInput.DEFAULT_INPUT_MODE;
 
 /*
@@ -32,6 +34,7 @@ public final class FileInputArray
   private final String name;
   private final String description;
   private FileInputMode inputMode;
+  private final JsonObject notes; // metadata as json
   private final List<String> sourceUrls;
   private final String targetDir;
 
@@ -45,16 +48,18 @@ public final class FileInputArray
     name = null;
     description = null;
     inputMode = DEFAULT_INPUT_MODE;
+    notes = DEFAULT_NOTES;
     sourceUrls = null;
     targetDir = null;
   }
 
-  public FileInputArray(String name1, String description1, FileInputMode inputMode1, List<String> sourceUrls1,
-                        String targetDir1)
+  public FileInputArray(String name1, String description1, FileInputMode inputMode1, JsonObject notes1,
+                        List<String> sourceUrls1, String targetDir1)
   {
     name = name1;
     description = description1;
     inputMode = (inputMode1 == null) ? DEFAULT_INPUT_MODE : inputMode1;
+    notes = (notes1 == null) ? DEFAULT_NOTES : notes1;
     sourceUrls = sourceUrls1;
     targetDir = targetDir1;
   }
@@ -65,6 +70,7 @@ public final class FileInputArray
   public String getName() { return name; }
   public String getDescription() { return description; }
   public FileInputMode getInputMode() { return inputMode; }
+  public JsonObject getNotes() { return notes; }
   public List<String> getSourceUrls() { return (sourceUrls == null) ? null : new ArrayList<>(sourceUrls); }
   public String getTargetDir() { return targetDir; }