+++ CHANGELOG.md
@@ -8,6 +8,15 @@ https://tapis.readthedocs.io/en/latest/technical/apps.html
 You may also reference live-docs based on the openapi specification here:
 https://tapis-project.github.io/live-docs
 
+## 1.4.1 - 2023-??-??
+
+Incremental improvements.
+
+### New features:
+- Add support for attribute *notes* for *fileInput* and *fileInputArray*.
+- Add support for *logConfig* in *parameterSet*. For supporting flexible redirection of stdout and stderr. 
+
+---------------------------------------------------------------------------
 ## 1.4.0 - 2023-07-07
 
 Incremental improvements.